+++ test/sdk.test.ts
@@ -440,12 +440,25 @@ describe('Tests the SDK functionality', () => {
     it('adds screenshot integration', () => {
       init({
         attachScreenshot: true,
+        enabled: true,
       });
 
       const actualOptions = mockedInitAndBind.mock.calls[0][secondArg] as ReactNativeClientOptions;
       const actualIntegrations = actualOptions.integrations;
 
       expect(actualIntegrations).toEqual(expect.arrayContaining([expect.objectContaining({ name: 'Screenshot' })]));
+    });
+
+    it('Should not add screenshot integration when the SDK is disabled', () => {
+      init({
+        attachScreenshot: true,
+        enabled: false,
+      });
+
+      const actualOptions = mockedInitAndBind.mock.calls[0][secondArg] as ReactNativeClientOptions;
+      const actualIntegrations = actualOptions.integrations;
+
+      expect(actualIntegrations).not.toEqual(expect.arrayContaining([expect.objectContaining({ name: 'Screenshot' })]));
     });
 
     it('no view hierarchy integration by default', () => {